Impl redirect option
It may be useful to send the user a separate file within the server --
perhaps a stylized .html file with embedded css, or a text file with
more extrapolation on the ban.

Chronologically,
1. Check if a redirect option is defined
2. Apply statusCode
3. Send the file, which inherently terminates

This circumvents the application of errorMessage.

res.sendFile is used instead of res.redirect...
GET '/' -> Banned? -> (redirect) GET '/403.html' -> Banned? -> ...
res.sendFile fixes this by terminating as soon as the file is sent.

res.sendFile requires an absolute path, so to specify a redirect,
app.use(ipfilter(blacklist, {
    log: false,
    redirect: __dirname + '/public/403.html',
    errorCode: 403
}));

Note, ipfilter.js.map is untouched.